requirements.txt
flask
requests
colorama
python-nmap
mysql-connector-python
